Add files via upload
<h1> AnyBooks-ZA </h1>

AnyBooks-ZA is a web application that consumes a GoogleBooks API to return Book cards that contain book information for any books hoosted in their library, some books are fully accessible and some do have payment restrictions. It has a built in reader for books that do not have any restrictions and the whole book is viewable and readable from within your browser. The book opens up in a new Tab.

This web app was built using HTML, CSS, jQuery and JavaScript.